+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/metrics/dns_metrics.py
@@ -13,9 +13,11 @@
 # limitations under the License.
 
 
+from typing import Final
+
 from opentelemetry.metrics import Histogram, Meter
 
-DNS_LOOKUP_DURATION = "dns.lookup.duration"
+DNS_LOOKUP_DURATION: Final = "dns.lookup.duration"
 """
 Measures the time taken to perform a DNS lookup
 Instrument: histogram